LayoutTests/accessibility/mac/input-replacevalue-userinfo-expected.txt


This tests value change notifications user info data when replacing the contents of an input field.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeEdit
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextChangeValue"] is "0"
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextEditType"] is AXTextEditTypeDelete
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeEdit
PASS results[resultIndex]["AXTextChangeValues"][1]["AXTextChangeValue"] is "1"
PASS results[resultIndex]["AXTextChangeValues"][1]["AXTextEditType"] is AXTextEditTypeInsert
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeEdit
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextChangeValue"] is " "
PASS results[resultIndex]["AXTextChangeValues"][0]["AXTextEditType"] is AXTextEditTypeInsert
PASS successfullyParsed is true

TEST COMPLETE
